server/requirements.txt
flask
flask-SQLAlchemy
flask-cors
flask_sqlalchemy
python-dotenv
requests
mysql-connector-python